+++ src/main/java/io/github/mzmine/modules/visualization/ims/imsVisualizer/RetentionTimeIntensityPlot.java
@@ -18,79 +18,85 @@
 
 package io.github.mzmine.modules.visualization.ims.imsVisualizer;
 
-import io.github.mzmine.datamodel.MassList;
-import io.github.mzmine.datamodel.RawDataFile;
-import io.github.mzmine.datamodel.Scan;
 import io.github.mzmine.gui.chartbasics.chartthemes.EStandardChartTheme;
 import io.github.mzmine.gui.chartbasics.gui.javafx.EChartViewer;
 import io.github.mzmine.main.MZmineCore;
-import io.github.mzmine.modules.visualization.ims.ImsVisualizerParameters;
-import io.github.mzmine.modules.visualization.spectra.simplespectra.SpectraVisualizerWindow;
-import io.github.mzmine.modules.visualization.spectra.simplespectra.datasets.MassListDataSet;
-import io.github.mzmine.parameters.ParameterSet;
-import io.github.mzmine.util.color.SimpleColorPalette;
-import javafx.application.Platform;
+import io.github.mzmine.modules.visualization.ims.ImsVisualizerTask;
 import org.jfree.chart.ChartFactory;
 import org.jfree.chart.JFreeChart;
-import org.jfree.chart.axis.AxisLocation;
 import org.jfree.chart.block.BlockBorder;
 import org.jfree.chart.entity.ChartEntity;
 import org.jfree.chart.entity.XYItemEntity;
 import org.jfree.chart.fx.interaction.ChartMouseEventFX;
 import org.jfree.chart.fx.interaction.ChartMouseListenerFX;
 import org.jfree.chart.plot.PlotOrientation;
+import org.jfree.chart.plot.ValueMarker;
 import org.jfree.chart.plot.XYPlot;
 import org.jfree.chart.renderer.xy.XYLineAndShapeRenderer;
+import org.jfree.chart.title.PaintScaleLegend;
+import org.jfree.chart.ui.RectangleAnchor;
+import org.jfree.chart.ui.TextAnchor;
 import org.jfree.data.xy.XYDataset;
 
 import java.awt.*;
 import java.util.logging.Logger;
 
-public class MobilityIntensityPlot extends EChartViewer {
+public class RetentionTimeIntensityPlot extends EChartViewer {
 
   private XYPlot plot;
   private JFreeChart chart;
   private Logger logger = Logger.getLogger(this.getClass().getName());
   static final Font legendFont = new Font("SansSerif", Font.PLAIN, 12);
+  private double selectedRetention;
+  private ValueMarker marker;
   private EStandardChartTheme theme;
-  private Scan selectedMobilityScan;
-  private Scan scans[];
-  private RawDataFile dataFiles[];
 
-  public MobilityIntensityPlot(XYDataset dataset, ParameterSet parameters) {
+  public RetentionTimeIntensityPlot(
+      XYDataset dataset,
+      ImsVisualizerTask imsVisualizerTask,
+      RetentionTimeMobilityHeatMapPlot retentionTimeMobilityHeatMapPlot) {
+
     super(
         ChartFactory.createXYLineChart(
-            "", "intensity", "", dataset, PlotOrientation.VERTICAL, true, true, false));
+            "",
+            "retention time",
+            "intensity",
+            dataset,
+            PlotOrientation.VERTICAL,
+            true,
+            true,
+            false));
     chart = getChart();
-    plot = chart.getXYPlot();
     theme = MZmineCore.getConfiguration().getDefaultChartTheme();
     theme.apply(chart);
-    dataFiles =
-        parameters
-            .getParameter(ImsVisualizerParameters.dataFiles)
-            .getValue()
-            .getMatchingRawDataFiles();
-    scans =
-        parameters
-            .getParameter(ImsVisualizerParameters.scanSelection)
-            .getValue()
-            .getMatchingScans(dataFiles[0]);
-
-    var renderer = new XYLineAndShapeRenderer();
+    plot = chart.getXYPlot();
+    this.selectedRetention = imsVisualizerTask.getSelectedRetentionTime();
+    var renderer = new XYLineAndShapeRenderer(true, true);
     renderer.setSeriesPaint(0, Color.GREEN);
-    renderer.setSeriesStroke(0, new BasicStroke(1.0f));
     renderer.setSeriesShapesVisible(0, false);
+    renderer.setSeriesStroke(0, new BasicStroke(1.0f));
 
     plot.setRenderer(renderer);
     plot.setBackgroundPaint(Color.WHITE);
     plot.setRangeGridlinePaint(Color.WHITE);
     plot.setDomainGridlinePaint(Color.WHITE);
-    plot.setRangeAxisLocation(AxisLocation.TOP_OR_RIGHT);
-    plot.getDomainAxis().setInverted(true);
-    plot.getRangeAxis().setVisible(false);
     chart.getLegend().setFrame(BlockBorder.NONE);
 
-    // mouse listener.
+    plot.clearDomainMarkers();
+    marker = new ValueMarker(selectedRetention);
+    marker.setPaint(Color.red);
+    marker.setLabelFont(legendFont);
+    marker.setStroke(new BasicStroke(2));
+    marker.setLabelAnchor(RectangleAnchor.TOP_RIGHT);
+    marker.setLabelTextAnchor(TextAnchor.BASELINE_CENTER);
+    plot.addDomainMarker(marker);
+
+    imsVisualizerTask.setSelectedRetentionTime(selectedRetention);
+    imsVisualizerTask.updateMobilityGroup();
+    //  marker to the mobility-retention time heatmap plot.
+    retentionTimeMobilityHeatMapPlot.getPlot().clearDomainMarkers();
+    retentionTimeMobilityHeatMapPlot.getPlot().addDomainMarker(marker);
+
     addChartMouseListener(
         new ChartMouseListenerFX() {
           @Override
@@ -100,14 +106,24 @@ public class MobilityIntensityPlot extends EChartViewer {
               XYItemEntity entity = (XYItemEntity) chartEntity;
               int serindex = entity.getSeriesIndex();
               int itemindex = entity.getItem();
-              double mobility = dataset.getYValue(serindex, itemindex);
-              for (int i = 0; i < scans.length; i++) {
-                if (scans[i].getMobility() == mobility) {
-                  selectedMobilityScan = scans[i];
-                  break;
-                }
-              }
-              updateChart();
+              selectedRetention = dataset.getXValue(serindex, itemindex);
+              // Get controller
+              imsVisualizerTask.setSelectedRetentionTime(selectedRetention);
+              imsVisualizerTask.updateMobilityGroup();
+
+              // setting the marker at seleted range.
+              plot.clearDomainMarkers();
+              marker = new ValueMarker(selectedRetention);
+              marker.setPaint(Color.red);
+              marker.setLabelFont(legendFont);
+              marker.setStroke(new BasicStroke(2));
+              marker.setLabelAnchor(RectangleAnchor.TOP_RIGHT);
+              marker.setLabelTextAnchor(TextAnchor.BASELINE_CENTER);
+              plot.addDomainMarker(marker);
+
+              //  marker to the mobility-retention time heatmap plot.
+              retentionTimeMobilityHeatMapPlot.getPlot().clearDomainMarkers();
+              retentionTimeMobilityHeatMapPlot.getPlot().addDomainMarker(marker);
             }
           }
 
@@ -116,27 +132,7 @@ public class MobilityIntensityPlot extends EChartViewer {
         });
   }
 
-  void updateChart() {
-    SpectraVisualizerWindow spectraWindow = new SpectraVisualizerWindow(dataFiles[0]);
-    spectraWindow.loadRawData(selectedMobilityScan);
-
-    // set colors depending on vision
-    SimpleColorPalette palette = MZmineCore.getConfiguration().getDefaultColorPalette();
-    Color posColor = palette.getPositiveColorAWT();
-    Color negColor = palette.getNegativeColorAWT();
-
-    // set color
-    XYPlot plotSpectra = (XYPlot) spectraWindow.getSpectrumPlot().getChart().getPlot();
-
-    // set color
-    plotSpectra.getRenderer().setSeriesPaint(0, posColor);
-
-    // add mass list
-    MassList[] massLists = selectedMobilityScan.getMassLists();
-    for (MassList massList : massLists) {
-      MassListDataSet dataset = new MassListDataSet(massList);
-      spectraWindow.getSpectrumPlot().addDataSet(dataset, negColor, true);
-    }
-    Platform.runLater(() -> spectraWindow.show());
+  public void setLegend(PaintScaleLegend legend) {
+    chart.addSubtitle(legend);
   }
 }
